pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wod-fusion"
version = "0.1.0"
description = "Waymo Perception Fusion to Ego Action"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Fusion2Drive Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "einops>=0.7.0",
    "timm>=1.0.0",  # For EfficientNet and vision backbones
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
]

[project.optional-dependencies]
waymo = [
    "waymo-open-dataset-tf-2-12-0>=1.6.0",
    "tensorflow>=2.12.0,<2.16.0",
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
]
export = [
    "onnx>=1.14.0",
    "onnxruntime>=1.16.0",
    "coremltools>=7.0",
]
carla = []  # Install CARLA PythonAPI separately
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
all = [
    "wod-fusion[waymo,export,dev]",
]

[project.scripts]
wod-train = "scripts.train:main"
wod-eval = "scripts.evaluate:main"
wod-demo = "scripts.demo:main"

[tool.hatch.build.targets.wheel]
packages = ["wod_fusion"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.4.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
